Adding the removal of unused links to pdf files to debian/rules.
These links are not useful anymore. They are created for each language but since
they have not got the language extension in their name they are constantly
overwritten after each group of pdf files is built.

For example, when the English pdf files are being built these links point to the
English pdf files, then when the Spanish pdf files are being built the links are
overwritten to point to the Spanish files and so on and so forth. In the end
they always link to the last pdf files of the last language. As of now, it it
Romanian.

These links end up in the final .deb packages unnecessarily, so, it is better to
remove them.