Reinstating codenames in terms.
Some codenames were removed from the terms section to make it easier to maintain
since codenames change in each release cycle. Now the manual takes advantage of
sisu's substitute command and it will be easier to maintain.

Instead of writing the codename for the stable or testing releases it is enough
to write {stable} or {testing} preceded by $ instead. Sisu will take charge of
substituting them for the actual names. The actual codenames are specified in
the file manual/en/live-manual.ssm.

In respect to sid this is irrelevant since the codename never changes.