Explicitly naming completely translated languages using the two letter code.
As the number of translated languages grows, the more useful it will be to have
information on which ones are 100% complete and which ones are not (yet).

This will be easier once the script analyzes 'all' the languages in one go and
outputs which ones are ready to go. This part is not implemented yet, but it
will, soon.