Untracking po4a configuration file.
po4a.cfg is a disposable configuration file which is deleted an recreated by the
Makefile after each run. So any changes to the configuration file have to be
tracked in the corresponding Makefile. Not in the file itself.

Besides, removing the file from the git repo will also allow experienced users
to use a modified configuration file if they wish to do so. It has to be added
to the .gitignore file, though, which is the next step.